Switches build process to make composite build locally possible.
For this there are new gradle task "enableCompositeBuild" and "disableCompositeBuild".
After running one of the tasks you have to reimport the gradle projects (in IDEA use the symbol the two arrows) to get the changes.

With the composite build enabled the libraries are build locally by using your local copy of the dependencies.
Important: Never commit the .composite-enable file. Otherwise, the travis builds will no longer work.